Package setup and org installation before processing src blocks.
The Org mode can be subtly broken if on old version is overridden
with a new one during one Emacs session. Unfortunately this was the
case when init.el was processed with the Org mode shipped by Emacs
and later installed a more recent version. This is a hotfix - a
real solution might be a way to process the org-src blocks without
actually using Org.